+++ modules/load_music.rb
